refactor(cli): split output.rs (1172 lines) into output/ modules
Split per-domain (json, text, snapshot, logs, network, storage, forms,
watch, diff, windows, record) to keep each file under the 150-line cap
defined in CLAUDE.md. Behavior preserved — all existing tests pass plus
new char tests in tests/output_smoke.rs (issue #70).
